data_ingestion/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pdfplumber==0.11.0
sentence-transformers==2.5.1
qdrant-client==1.7.3
tqdm==4.66.2
numpy==1.26.4
pandas==2.2.1
